backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
alembic>=1.13
pydantic[email]>=2.0
email-validator>=2.0.0
pydantic-settings>=2.0
bcrypt>=4.0
PyJWT>=2.8
cryptography>=42.0
celery[redis]>=5.4
redis>=5.0
boto3>=1.34
slowapi>=0.1.9
httpx>=0.27
python-multipart>=0.0.9
pytest>=8.0
pytest-asyncio>=0.24
pytest-cov>=5.0
anyio>=4.0
aiosqlite>=0.20
ruff>=0.5

# Phase 1 — AI/media services
edge-tts>=6.1
deep-translator>=1.11
yt-dlp>=2024.0
google-generativeai>=0.8
Pillow>=10.0
# openai-whisper>=20231117  # Install separately — requires torch
